Update konnected-contact-sensor.groovy
Cleaned up debug, changed tile type to contact (standard), changed tile name to contactDisplay, added a preferences section for dynamically updating the labels, and a routine for the same.   Works fine but already installed devices will have to run through the preferences on each and add labels.  There is probably a way to say if the preferences are blank then they equal the default Open and Closed but I didn't figure that out.  Or maybe make a separate "Generic Binary Input" device.   Idk...use what you can from it.