Update konnected-motion-sensor.groovy
Is there any reason not to use a plain if then statement for state from parent instead of the larger case statement?  It would take care of your default value if you don't get a 0 or 1 passed down and looks a little cleaner I would think.  Also change the default tile name to same as capability