Update konnected-contact-sensor.groovy
Same as last commit except do not flip background image (keep actual status of open and closed for the icon) but still flip color and also added a label for "Armed"/"Disarmed".  

Personally this will be less confusing since the icon/background image would be consistent to the actual status of the input but still allow the color change.  I had both ways in my device handlers and after a few days I felt I rather know the actual underlying status (with the icon).  Of course if you change the icon it's not going to matter but still.