Fix OTA updates when connected to AWS.
This resolves an out of memory issue when updating via Over-The-Air while connected to a secure MQTT broker/AWS.
The process creates a file and writes the update info into it. The device is then rebooted and the file, if detected
after connection to WiFi, is read from and the update performed using the data in the file.